+++ ImageFeed/AppDelegate.swift
@@ -20,11 +20,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     
     func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
         let sceneConfiguration = UISceneConfiguration(
-            name: "Main", sessionRole: connectingSceneSession.role)
+            name: "Main",
+            sessionRole: connectingSceneSession.role
+        )
         sceneConfiguration.delegateClass = SceneDelegate.self
         return sceneConfiguration
-        
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
     }
     
     func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
