+++ ImageFeed/Controllers/Auth/WebViewViewController.swift
@@ -51,11 +51,11 @@ final class WebViewViewController: UIViewController {
         super.viewWillAppear(animated)
         estimatedProgressObservation = webView.observe(
             \.estimatedProgress,
-             options: [],
-             changeHandler: { [weak self] _, _ in
+             options: []
+        ) { [weak self] _, _ in
                  guard let self = self else { return }
                  self.updateProgress()
-             })
+             }
         updateProgress()
     }
