+++ ImageFeed/Models/Like.swift
@@ -1,0 +1,12 @@
+//
+//  Like.swift
+//  ImageFeed
+//
+//  Created by Nataliya TRUFANOVA on 05/01/2024.
+//
+
+import Foundation
+
+struct Like: Codable {
+    let photo: PhotoResult?
+}
