+++ ImageFeed/Models/UrlsResult.swift
@@ -1,0 +1,14 @@
+//
+//  UrlsResult.swift
+//  ImageFeed
+//
+//  Created by Nataliya TRUFANOVA on 05/01/2024.
+//
+
+import Foundation
+
+struct UrlsResult: Codable {
+    let full: String
+    let thumb: String
+    let small: String
+}
